[Game] Move game class to new file
