+++ src/client/client_engine.cpp
@@ -1,7 +1,7 @@
 #include "client_engine.h"
 
-#include "client.h"
 #include "client_config.h"
+#include "game.h"
 #include "gl/framebuffer.h"
 #include "gl/gl_errors.h"
 #include "gl/primitive.h"
@@ -17,10 +17,6 @@
 #include <SFML/System/Clock.hpp>
 #include <common/scripting/script_engine.h>
 #include <glad/glad.h>
-#include <memory>
-#include <thread>
-
-#include <server_engine.h>
 
 namespace {
     struct FPSCounter final {
@@ -40,98 +36,7 @@ namespace {
         }
     };
 
-    class Game {
-      public:
-        bool initGame(ClientConfig config, const std::string& ipAddress);
-        bool initGame(ClientConfig config);
-
-        void stopGame();
-
-        // Ran during game loop
-        void onMouseRelease(sf::Mouse::Button button);
-        void input(sf::Window& window, const Keyboard& keyboard,
-                   const InputState& inputState);
-        void update(float dt);
-        void render();
-
-      private:
-        bool init(ClientConfig config);
-
-        std::unique_ptr<Client> m_client;
-        std::unique_ptr<ServerLauncher> m_serverLauncher;
-    };
-
     //
-    bool Game::initGame(ClientConfig config)
-    {
-        config.serverIp = LOCAL_HOST;
-
-        m_serverLauncher =
-            std::make_unique<ServerLauncher>(ServerConfig{8, 8}, sf::milliseconds(100));
-        m_serverLauncher->runAsThread();
-
-        return init(config);
-    }
-
-    bool Game::initGame(ClientConfig config, const std::string& ipAddress)
-    {
-        config.serverIp = ipAddress;
-
-        return init(config);
-    }
-
-    bool Game::init(ClientConfig config)
-    {
-        m_client = std::make_unique<Client>();
-        if (!m_client->init(config,
-                            (float)config.windowWidth / (float)config.windowHeight)) {
-            stopGame();
-            return false;
-        }
-        return true;
-    }
-
-    void Game::stopGame()
-    {
-        if (m_serverLauncher) {
-            m_serverLauncher->stop();
-            m_serverLauncher.release();
-        }
-        if (m_client) {
-            m_client->endGame();
-            m_client->destroy();
-            m_client.release();
-        }
-    }
-
-    void Game::onMouseRelease(sf::Mouse::Button button)
-    {
-        if (m_client) {
-            m_client->onMouseRelease(button);
-        }
-    }
-
-    void Game::input(sf::Window& window, const Keyboard& keyboard,
-                     const InputState& inputState)
-    {
-        if (m_client) {
-            m_client->handleInput(window, keyboard, inputState);
-        }
-    }
-
-    void Game::update(float dt)
-    {
-        if (m_client) {
-            m_client->update(dt);
-        }
-    }
-
-    void Game::render()
-    {
-        if (m_client) {
-            m_client->render();
-        }
-    }
 } // namespace
 
 void runClientEngine(const ClientConfig& config)
